Initial work on rearranging the directory structure and regression testing

+++ OMeet/si_stick_finish.php
@@ -3,7 +3,7 @@
 // Assume that this is already included - otherwise bad things will happen
 //require "common_routines.php";
 
-function record_finish_by_si_stick($event, $si_results_string) {
+function record_finish_by_si_stick($event, $key, $si_results_string) {
   //echo "Looking at {$si_results_string}\n";
   $return_info = array();
 
@@ -31,7 +31,7 @@ function record_finish_by_si_stick($event, $si_results_string) {
   $return_info["si_stick"] = $si_stick;
 
   // Find the competitor with this si_stick
-  $competitor_directory = "./{$event}/Competitors";
+  $competitor_directory = get_competitor_directory($event, $key, ".."); 
   $competitor_list = scandir("{$competitor_directory}");
   $competitor_list = array_diff($competitor_list, array(".", ".."));
 
@@ -40,7 +40,7 @@ function record_finish_by_si_stick($event, $si_results_string) {
       $competitor_stick = file_get_contents("{$competitor_directory}/{$competitor}/si_stick");
       if ($competitor_stick == $si_stick) {
         if (!file_exists("{$competitor_directory}/{$competitor}/controls_found/start")) {
-          $save_error = validate_and_save_results($event, $competitor, $si_stick, $start_pieces, $finish_pieces, $result_pieces);
+          $save_error = validate_and_save_results($event, $key, $competitor, $si_stick, $start_pieces, $finish_pieces, $result_pieces);
           if ($save_error == "") {
             $return_info["competitor_id"] = $competitor;
             $return_info["course"] = file_get_contents("{$competitor_directory}/{$competitor}/course");
@@ -70,19 +70,20 @@ function record_finish_by_si_stick($event, $si_results_string) {
   return($return_info);
 }
 
-function validate_and_save_results($event, $competitor, $si_stick, $start_pieces, $finish_pieces, $result_pieces) {
+function validate_and_save_results($event, $key, $competitor, $si_stick, $start_pieces, $finish_pieces, $result_pieces) {
   global $TYPE_FIELD, $SCORE_O_COURSE;
 
-  $competitor_path = "./{$event}/Competitors/{$competitor}";
+  $competitor_path = get_competitor_path($competitor, $event, $key, "..");
   if (!file_exists($competitor_path) || file_exists("{$competitor_path}/controls_found/start") || ($si_stick != file_get_contents("{$competitor_path}/si_stick"))) {
     // Houston, we have a problem
     return ("Competitor {$competitor} does not seem to match stick {$si_stick}");
   }
 
   $course = file_get_contents("{$competitor_path}/course");
-  $controls_info = read_controls("./{$event}/Courses/{$course}/controls.txt");
+  $courses_path = get_courses_path($event, $key, "..");
+  $controls_info = read_controls("{$courses_path}/{$course}/controls.txt");
   $controls_on_course = array_map(function ($elt) { return($elt[0]); }, $controls_info);
-  $course_properties = get_course_properties("./{$event}/Courses/{$course}");
+  $course_properties = get_course_properties("{$courses_path}/{$course}");
   $is_score_course = (isset($course_properties[$TYPE_FIELD]) && ($course_properties[$TYPE_FIELD] == $SCORE_O_COURSE));
 
 
@@ -120,7 +121,15 @@ function validate_and_save_results($event, $competitor, $si_stick, $start_pieces
 
   // Do this last so that it can be used to determine if the results were saved properly
   // Do NOT save the finish time, as the finish_course will do that
-  file_put_contents("{$competitor_path}/controls_found/start", $start_pieces[1]);
+  if (file_exists("{$competitor_path}/mass_si_stick_start")) {
+    if ($start_pieces[1] != 0) {
+      // HMMMM, this shouldn't happen, error somehow
+    }
+    file_put_contents("{$competitor_path}/controls_found/start", file_get_contents("{$competitor_path}/mass_si_stick_start"));
+  }
+  else {
+    file_put_contents("{$competitor_path}/controls_found/start", $start_pieces[1]);
+  }
 
   return("");
 }
